timestring generation externalization
