+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "add-to-calendar-button",
-  "version": "2.5.0-next.5",
+  "version": "2.5.0-next.6",
   "engines": {
     "node": ">=16.18.1",
     "npm": ">=8.19.2"
